requirements.txt
# MCP Core
mcp>=1.0.0

# HTTP Requests
aiohttp>=3.8.0
requests>=2.28.0

# Telegram
python-telegram-bot>=20.0

# RSS Feed Parsing
feedparser>=6.0.0

# Email
Jinja2>=3.1.0

# Date Processing
python-dateutil>=2.8.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Optional: Advanced NLP (for future enhancement)
# spacy>=3.4.0
# transformers>=4.20.0

# Optional: Database (for future enhancement)  
# sqlalchemy>=1.4.0
# sqlite3